+++ src/hooks/login/helpers/index.ts
@@ -1,1 +1,2 @@
 export * from './clearInitiatedLogins';
+export * from './initAndValidateNativeAuthToken';